Spindle improvements (wip)
- export functionality moved to new table component

#3256
