+++ waltz-ng/client/measurable-relationship/components/panel/related-measurables-panel.js
@@ -19,11 +19,9 @@
 
 import _ from "lodash";
 import {initialiseData} from "../../../common";
-import {downloadTextFile} from "../../../common/file-utils";
 import {sameRef} from "../../../common/entity-utils";
 import {CORE_API} from "../../../common/services/core-api-utils";
 import {entity} from "../../../common/services/enums/entity";
-import {relationshipKind} from "../../../common/services/enums/relationship-kind";
 import {getEnumName} from "../../../common/services/enums";
 import {sanitizeRelationships} from "../../measurable-relationship-utils";
 
@@ -115,48 +113,6 @@ function mkGridData(ref,
         .filter(r => r !== null)
         .sortBy(['a.name', 'b.name'])
         .value()
-}
-
-
-function mkExportData(relationships = [], categories = [], measurables = []) {
-    const categoriesById = _.keyBy(categories, 'id');
-    const categoriesByMeasurableId = _.chain(measurables)
-        .map(m => ({
-            measurableId: m.id,
-            category: categoriesById[m.categoryId]
-        }))
-        .keyBy('measurableId')
-        .value();
-
-    const getType = (id, kind) => {
-        return kind === 'MEASURABLE'
-            ? categoriesByMeasurableId[id].category.name
-            : getEnumName(entity, kind);
-    };
-
-    const columnNames = [[
-        "From",
-        "From type",
-        "To",
-        "To type",
-        "Relationship",
-        "Description",
-        "Last Updated At",
-        "Last Updated By"
-    ]];
-
-    const exportData = _.map(relationships, r => [
-        r.a.name,
-        getType(r.a.id, r.a.kind),
-        r.b.name,
-        getType(r.b.id, r.b.kind),
-        getEnumName(relationshipKind, r.relationship),
-        r.description,
-        r.lastUpdatedAt,
-        r.lastUpdatedBy
-    ]);
-
-    return columnNames.concat(exportData);
 }
 
 
@@ -253,11 +209,6 @@ function controller($q, $timeout, serviceBroker, notification) {
     };
 
     vm.selectionFilterFn = DEFAULT_SELECTION_FILTER_FN;
-
-    vm.export = () => {
-        const data = mkExportData(vm.relationships, vm.categories, vm.measurables);
-        downloadTextFile(data, ",", "related_viewpoints.csv");
-    };
 
 
     // -- API --
